# [xy] Fix data loaders and improve performance