# Summary
<!-- Brief summary of what your code does -->
This PR includes a couple of fixes and improves performance
* Fix empty response from Snowflake loader execute method

For a dataset with 10 million rows
* Speed up column type detection. Reduce summary overview chart plot time from 98s to 2s.
* Speed up histogram calculation. Reduce time from 7.3s to 1.2s

# Tests
<!-- How did you test your change? -->
tested locally

cc:
<!-- Optionally mention someone to let them know about this pull request -->